.do: update

+++ .do/do.sh
@@ -1,10 +1,10 @@
 #!/usr/bin/env bash
 #
 
-if [[ $1 == '-h' ]]; then
+if [[ $* =~ (^|[[:space:]])-h([[:space:]]|$) ]]; then
   echo "
   Example:
-      ${0##*/} headscale 0.17.0 ~/Git/gopkg-vendors
+      ${0##*/} ~/Git/gopkg-vendors headscale 0.17.0
 "
   exit
 fi
@@ -13,9 +13,9 @@ _FIND_PATH="/var/cache/distfiles"
 
 _THIS_PATH=$(dirname $(realpath $0))
 
-_PKG_NAME=${1}
-_MOD_VER=${2}
-_VCS_DIR=${3}
+_VCS_DIR=${1}
+_PKG_NAME=${2}
+_MOD_VER=${3}
 
 _PKG_PATH=$(ls -1 ${_FIND_PATH%/}/${_PKG_NAME}-${_MOD_VER}.tar.*)
 [[ -z ${_PKG_PATH} ]] && \
@@ -36,4 +36,6 @@ _PKG_ARGS=${_PKG_ARGS//%MYDIR%/${_THIS_PATH}/${_PKG_NAME}}
 
 eval "${_THIS_PATH}/../gen.sh -b${_PKG_NAME} -u ${_PKG_ARGS} -p \"${_VCS_DIR}\" \"${_MOD_DIR}\" ${_MOD_VER}"
 
-rm -rf "${_TMPDIR}"
+if [[ -z ${4} ]]; then
+  rm -rf "${_TMPDIR}"
+fi
